requirements.txt
cryptography>=42.0.0
pyserial>=3.5
spidev>=3.6; platform_machine == "armv7l" or platform_machine == "aarch64"
RPi.GPIO>=0.7.1; platform_machine == "armv7l" or platform_machine == "aarch64"
click>=8.1.0
prompt_toolkit>=3.0.0
rich>=13.0.0
